feat: add a helm chart with the releaser action (#349)
* feat: Create an initial version of a helm chart

Currently, Helm is the only package manager for Kubernetes, and it's
very widely used across many Kubernetes users. So I've thought that it
could make sense to start distributing sql_exporter with helm

* build: Add a Github Workflow to manage helm chart

This action will lint the chart on every change. If chart was updated,
but its version was not, the action will be failed, that should make
the delivery more smooth.

If commit was pushed to the main branch, the action will release a new
chart version. It will create a new GitHub release and push to the
gh-pages branch. This branch has to be configured as the one for serving
GitHub Pages. Then it will start action as a Helm repository.

* refactor: Apply some formatting and update values