build: Compare current chart to latest released
I've added a step to the helm-workflow, that is adding an annotation to
the helm chart manifests. This annotations contains a SHA of a current
commit. If the action is triggered by puhsing to the master, this
annotations will be preserved and will make it to the upstream
manifests, though it's not visible in the git repo.

I'm adding to identify the the commit to which we should compare the
current version of the chart. If it's not accessible, the chart will be
compared to the previous commit, that basically means that there is no
released version yet.

I've also added a repo URL to chart annotations, so we can get the chart
from the upstream repo to retrieve the SHA annotations from the upstream
chart. It should make it possible to use the same workflow in forked
repos, if an intention behind forking is merging something to the
upstream repo.