build: Switch helm-releaser to the main branch
We need to use some features that are available only in the main branch
at the moment. So until the next version is released, I'd suggest
switching to the main revision.

Also, I've removed a notice about helm tests from NOTES.TXT, because we
are not rendering tests anymore by default, and hence this note might be
confusing